Fix NET-155
git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/net/branches/JDK_1_5_BRANCH@523750 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/examples/nntp/ExtendedNNTPOps.java
@@ -57,8 +57,8 @@ public class ExtendedNNTPOps {
 			// XOVER
 			NewsgroupInfo testGroup = new NewsgroupInfo();
 			client.selectNewsgroup("alt.test", testGroup);
-			int lowArticleNumber = testGroup.getFirstArticle();
-			int highArticleNumber = lowArticleNumber + 100;
+			long lowArticleNumber = testGroup.getFirstArticle();
+			long  highArticleNumber = lowArticleNumber + 100;
 			Article[] articles = NNTPUtils.getArticleInfo(client, lowArticleNumber, highArticleNumber);
 
 			for (int i = 0; i < articles.length; ++i) {
